requirements.txt
APScheduler==3.6.3
certifi==2023.7.22
cffi==1.14.4
cryptography==41.0.4
decorator==4.4.2
pycparser==2.20
python-telegram-bot==13.1
pytz==2020.5
six==1.15.0
tornado==6.3.3
tzlocal==2.1
wincertstore==0.2
tensorflow
python-telegram-bot>=20.0